+++ app/src/main/java/com/safelogj/lim/request/SearchChatRequest.java
@@ -1,4 +1,4 @@
 package com.safelogj.lim.request;
 
-public record SearchChatRequest (String username, String password, Long queryUserId) {
+public record SearchChatRequest (String username, String password, Integer queryUserId) {
 }
