+++ app/src/main/java/com/safelogj/lim/request/SendMessageRequest.java
@@ -1,7 +1,7 @@
 package com.safelogj.lim.request;
 
 
-public record SendMessageRequest(String username, String password, Long userId, Long chatId, String text, String type,
+public record SendMessageRequest(String username, String password, Integer userId, Integer chatId, String text, String type,
                                  String filePath, String fileName, String chatName) {
 
 }
