+++ limserver/src/main/java/com/safelogj/limserver/model/User.java
@@ -1,11 +1,10 @@
 package com.safelogj.limserver.model;
 
 public class User {
-    public long id;
+    public int id;
     public String username;
     public String displayName;
     public String publicKey;
     public String privateHash;
     public boolean isDeleted;
-
 }
