+++ limserver/src/main/java/com/safelogj/limserver/request/GetMessagesRequest.java
@@ -3,7 +3,7 @@ package com.safelogj.limserver.request;
 
 import java.util.List;
 
-public record GetMessagesRequest(String username, String password, Long lastMessageId, List<Long> interlocutorIds) {
+public record GetMessagesRequest(String username, String password, Long lastMessageId, List<Integer> interlocutorIds) {
 
     public boolean isValidRequest() {
         return lastMessageId != null
