+++ limserver/src/main/java/com/safelogj/limserver/request/HideChatRequest.java
@@ -1,6 +1,6 @@
 package com.safelogj.limserver.request;
 
-public record HideChatRequest (String username, String password, Long chatId) {
+public record HideChatRequest (String username, String password, Integer chatId) {
 
     public boolean isValidRequest() {
         return chatId != null
