+++ limserver/src/main/java/com/safelogj/limserver/request/SearchChatRequest.java
@@ -1,6 +1,6 @@
 package com.safelogj.limserver.request;
 
-public record SearchChatRequest (String username, String password, Long queryUserId) {
+public record SearchChatRequest (String username, String password, Integer queryUserId) {
 
     public boolean isValidRequest() {
         return username != null && !username.isEmpty()
